1. Last commit before replacing the climb sound thresholds with the simpler on     and off zones.
 2. The plan is to create two new pages; one that shows ground speed and one
    that shows only altitude. The last one should also feature altitude sound
    instead of climb sound. Since the location API (that will provide ground
    speed) uses double instead of float, a lot of functions now use double
    instead of float.

 3. The Averager class is now a template class that supports any number type.

 4. A lot of the Vario*-classes have changed names to Value* and some of them
    are supposed to be used only as base classes, like ValueAudio that is
    inherited by AltitudeAudio and ClimbAudio.

 5. VarioController have changed name to Controller but is put into a namespace
    called ColorVario, i.e. ColorVario::Controller.

 6. Averaging is no longer done in the Variometer class. Instead all classes
    that needs averaging have their own Averager. This enables different
    averaging times for different items.

 7. I am testing a few new icons. The grey disabled icons look good! The old
    icons are stored under doc/Icons/<version>.

 8. BluetoothLowEnergyAdvertiser is renamed to BluetoothAdvertiser. Shorter and
    ok since it is the only advertiser we need.

 9. Unit conversion in the labels are now handled by a callback that points to
    one of the conversion functions in UnitConversionFunctions.h.

10. MathFunctions.h replaces math.h and UnitConversionFunctions.h replaces
    unit.h.

11. Screenshots are reorganised under doc/Screenshots/<version>.

12. Some of the icon files are renamed.

+++ src/Display.cpp
@@ -27,16 +27,16 @@ display_state_e Kystsoft::Display::state() const
 	return state;
 }
 
-float Kystsoft::Display::brightness() const
+double Kystsoft::Display::brightness() const
 {
 	int ibrightness = 0;
 	int error = device_display_get_brightness(0, &ibrightness);
 	if (error != DEVICE_ERROR_NONE)
 		throw TizenError("device_display_get_brightness", error);
-	return float(ibrightness) / maxBrightness();
+	return double(ibrightness) / maxBrightness();
 }
 
-void Kystsoft::Display::setBrightness(float brightness)
+void Kystsoft::Display::setBrightness(double brightness)
 {
 	if (brightness < 0)
 		return;
@@ -53,7 +53,7 @@ void Kystsoft::Display::setBrightness(float brightness)
 			return;
 		}
 	}
-	int ibrightness = int(brightness * maxBrightness() + 0.5f);
+	int ibrightness = int(brightness * maxBrightness() + 0.5);
 	if (ibrightness < 1)
 		ibrightness = 1;
 	int error = device_display_set_brightness(0, ibrightness);
